+++ tests/snippets/test_ellipsoid_main_axes.py
@@ -17,7 +17,6 @@ print("A([1,2,3])     =", A(p))
 print("A(pt).vertices =", A(pt).vertices[0])
 
 maxes = A.compute_main_axes()
-
 arr1 = Arrow([0,0,0], maxes[0]).c('r', 0.5)
 arr2 = Arrow([0,0,0], maxes[1]).c('g', 0.5)
 arr3 = Arrow([0,0,0], maxes[2]).c('b', 0.5)
@@ -29,4 +28,5 @@ sphere2 = sphere1.clone().apply_transform(A)
 
 show([[sphere1, __doc__], 
       [sphere2, arr1, arr2, arr3, str(M)]],
-     N=2, axes=1, bg='bb')
+     N=2, axes=1, bg='bb').close()
+