+++ docs/changes.md
@@ -13,6 +13,7 @@
 - add `transformations.compute_main_axes()`
 - add `transformations.__call__()` to apply it
 - fix small bug in `pointcloud.distance_to()`
+- add `applications.MorphPlotter()` to morph a polygonal mesh to a target mesh
 
 
 ## Breaking changes
@@ -37,6 +38,7 @@
 
 ## New/Revised Examples
 ```
+examples/advanced/warp4c.py
 examples/volumetric/slab_vol.py
 examples/other/madcad1.py
 examples/other/tetgen1.py